reconciler: rename clusterworkspacedeletion -> workspacedeletion
